backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
bcrypt==3.2.2
python-jose[cryptography]
email-validator
PyJWT
# The logic: FastAPI is our engine, SQLAlchemy is our database bridge, passlib/bcrypt is our password protection shield, and python-jose is our famous JWT VIP ticket.
